==> 5_tract_profile_analysis.ipynb <==
# Create TractSeg tract profiles models
## This is only used in the supplement
### Start by importing packages and loading the dataframe

import pandas as pd
import numpy as np
import sklearn.preprocessing
df = pd.read_pickle('df.pkl')

### Text generated from here goes into "subjects.txt" (look at code repo or TractSeg documentation for more info)
#### Uncomment everything but the model you wish to run
#### Use the examples below to make your own models!

df_tractseg = df.copy()
factors_corr = ['TOWRE','SEX','HAND','ICV']
factors_group = ['SEX','HAND','AGE','gFA','TRRDxSEX']
# Demean and rescale all factors
factors_corr_normed = sklearn.preprocessing.scale(df_tractseg[factors_corr], with_mean=True, with_std=True, copy=True)
factors_group_normed = sklearn.preprocessing.scale(df_tractseg[factors_group], with_mean=True, with_std=True, copy=True)

# TOWRE CORRELATION ACROSS WHOLE COHORT
for sub,model in zip(df_tractseg['subjects'],factors_corr_normed):
    print(sub,model[0],model[1],model[2],model[3])

# FA GROUP DIFF - HIGH TOWRE VS. LOW TOWRE
#for sub,group,model in zip(df_tractseg['subjects'],df_tractseg['TR_RD'],factors_group_normed):
#    print(sub,group,model[0],model[1],model[2],model[3],model[4])
